moved string storage change report to its own file
